chore: Move 'Header' from pages to components
